Fixing various warnings in gio/xdgmime/xdgmimemagic.c
Note that line 464, there is a strange test that I believe to be always false.

gio/xdgmime/xdgmimemagic.c: In function ‘_xdg_mime_magic_parse_magic_line’:
gio/xdgmime/xdgmimemagic.c:406:18: error: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Werror=sign-compare]
   if (bytes_read != matchlet->value_length)
                  ^~
gio/xdgmime/xdgmimemagic.c:426:22: error: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Werror=sign-compare]
       if (bytes_read != matchlet->value_length)
                      ^~
gio/xdgmime/xdgmimemagic.c:464:34: error: comparison of integer expressions of different signedness: ‘unsigned int’ and ‘int’ [-Werror=sign-compare]
       if (matchlet->range_length == -1)
                                  ^~
gio/xdgmime/xdgmimemagic.c:488:18: error: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Werror=sign-compare]
    for (i = 0; i < matchlet->value_length; i = i + matchlet->word_size)
                  ^
gio/xdgmime/xdgmimemagic.c: In function ‘_xdg_mime_magic_matchlet_compare_to_data’:
gio/xdgmime/xdgmimemagic.c:526:32: error: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Werror=sign-compare]
   for (i = matchlet->offset; i < matchlet->offset + matchlet->range_length; i++)
                                ^
gio/xdgmime/xdgmimemagic.c:535:18: error: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Werror=sign-compare]
    for (j = 0; j < matchlet->value_length; j++)
                  ^
gio/xdgmime/xdgmimemagic.c:547:18: error: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Werror=sign-compare]
    for (j = 0; j <  matchlet->value_length; j++)
                  ^

+++ gio/xdgmime/xdgmimemagic.c
@@ -316,7 +316,7 @@ _xdg_mime_magic_parse_magic_line (FILE              *magic_file,
   int c;
   int end_of_file;
   int indent = 0;
-  int bytes_read;
+  size_t bytes_read;
 
   assert (magic_file != NULL);
 
@@ -460,8 +460,9 @@ _xdg_mime_magic_parse_magic_line (FILE              *magic_file,
 	{
 	  _xdg_mime_magic_matchlet_free (matchlet);
 	  return XDG_MIME_MAGIC_EOF;
-	}
-      if (matchlet->range_length == -1)
+        }
+      /* FIXME: Is this test really needed ? */
+      if (matchlet->range_length == (unsigned int) -1)
 	{
 	  _xdg_mime_magic_matchlet_free (matchlet);
 	  return XDG_MIME_MAGIC_ERROR;
@@ -476,7 +477,7 @@ _xdg_mime_magic_parse_magic_line (FILE              *magic_file,
       if (matchlet->word_size > 1)
 	{
 #if LITTLE_ENDIAN
-	  int i;
+	  size_t i;
 #endif
 	  if (matchlet->value_length % matchlet->word_size != 0)
 	    {
@@ -522,7 +523,7 @@ _xdg_mime_magic_matchlet_compare_to_data (XdgMimeMagicMatchlet *matchlet,
 					  const void           *data,
 					  size_t                len)
 {
-  int i, j;
+  size_t i, j;
   for (i = matchlet->offset; i < matchlet->offset + matchlet->range_length; i++)
     {
       int valid_matchlet = TRUE;
